better-eng: reconstructed project structure
