+++ src/lib/banned_prompts.py
@@ -2,14 +2,16 @@ import asyncio
 
 import aiofiles
 
+import settings
+
 
 async def loads_banned_words():
     prompt = set()
-    filename = "banned_words.txt"
-    async with aiofiles.open(filename, "r") as r:
+    async with aiofiles.open(settings.BANNED_PROMPTS_FILE_PATH, "r") as r:
         for line in await r.readlines():
             prompt.add(line.strip())
-
+    
     return prompt
 
-BANNED_PROMPT = asyncio.run(loads_banned_words())
+
+BANNED_PROMPTS = asyncio.run(loads_banned_words())
