+++ src/lib/exceptions.py
